LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scrollable-overflow-padding-expected.txt


PASS scrollable-container 1
FAIL scrollable-container 2 assert_equals:
<scrollable-container data-expected-scroll-width="260">
    <div style="width: 200px; height: 0; margin-right: 50px; left: -1000px;"></div>
  </scrollable-container>
scrollWidth expected 260 but got 95
PASS scrollable-container 3
PASS scrollable-container 4
PASS scrollable-container 5
PASS scrollable-container 6
FAIL scrollable-container 7 assert_equals:
<scrollable-container data-expected-scroll-width="260">
    <div style="width: 200px; height: 0; margin-left: 50px; right: -1000px;"></div>
  </scrollable-container>
scrollWidth expected 260 but got 95
PASS scrollable-container 8
PASS scrollable-container 9
PASS scrollable-container 10
PASS scrollable-container 11
FAIL scrollable-container 12 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-bottom: 50px; top: -1000px;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 13
FAIL scrollable-container 14 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-bottom: 50px; top: -1000px; display: inline-block;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 15
PASS scrollable-container 16
FAIL scrollable-container 17 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-top: 50px; bottom: -1000px;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 18
FAIL scrollable-container 19 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-top: 50px; bottom: -1000px; display: inline-block;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 20
PASS scrollable-container 21
FAIL scrollable-container 22 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-bottom: 50px; top: -1000px;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 23
FAIL scrollable-container 24 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-bottom: 50px; top: -1000px; display: inline-block;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 25
PASS scrollable-container 26
FAIL scrollable-container 27 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-top: 50px; bottom: -1000px;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 28
FAIL scrollable-container 29 assert_equals:
<scrollable-container data-expected-scroll-height="270">
    <div style="width: 0; height: 200px; margin-top: 50px; bottom: -1000px; display: inline-block;"></div>
  </scrollable-container>
scrollHeight expected 270 but got 105
PASS scrollable-container 30
